requirements.txt
aiogram==3.2.0
telethon==1.34.0
google-api-python-client==2.110.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
python-dotenv==1.0.0
